Allow GRD reader to process a custom GRDP file stored in a folder different from its parent GRD file.
As of today, GRIT assumes that files referenced from the GRD file
are relative to a path stored in the root <grit> node. This change
fixes it by allowing grd reader to process a GRDP file stored in
an arbitrary path as long as the file exists.

BUG=729620
TEST=python tools/grit/grit.py unit
R=benrg@chromium.org,joi@chromium.org,thakis@chromium.org

Review-Url: https://codereview.chromium.org/2923103002
Cr-Original-Commit-Position: refs/heads/master@{#482829}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d93265c38767d0886c7bddcfaa643d8b3aa7c224